Potential fix for code scanning alert no. 12: Wrong type of arguments to formatting function

+++ fs/proc/pid.c
@@ -72,7 +72,7 @@ static int proc_pid_stat_show(struct proc_entry *entry, struct proc_data *buf) {
     // bunch of shit that can only be accessed by a debugger
     proc_printf(buf, "%lu ", 0l); // startcode
     proc_printf(buf, "%lu ", 0l); // endcode
-    proc_printf(buf, "%lu ", task->mm ? task->mm->stack_start : 0);
+    proc_printf(buf, "%lu ", (unsigned long) (task->mm ? task->mm->stack_start : 0ul));
     proc_printf(buf, "%lu ", 0l); // kstkesp
     proc_printf(buf, "%lu ", 0l); // kstkeip
